server/requirements.txt
anyio==4.6.2.post1
azure-storage-blob==12.10.0
bcrypt==4.2.1
firebase_admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask_JWT_Extended==4.7.1
groq==0.12.0
google_api_python_client==2.154.0
PyJWT==2.10.1
pymongo[tls]==4.10.1
PyMuPDF==1.24.14
pyparsing==3.2.0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2